- hidden id (input) - new login/new_task tests - SQL dump

+++ updateTask.php
@@ -1,29 +1,19 @@
 ﻿<?php
-require 'sessionCode/session.php';
 require 'dbCode/connect.php';
-$currentValue = current($_POST);
-switch ($currentValue){
-	case "Редактировать":
-		$selected = key($_POST);
-		$comment = showTask($selected);
+$selected = $_POST['id'];
 
-		$h2 = 'Редактировать задачу';
-		require 'header.php';
-		require 'templates/updateTask.php';		
-		require 'footer.html';
-		break;
-
-	case "Удалить задание":
-		header('Location: tasks.php');
-		$selected = key($_POST);
-		deleteTask($selected);
-		break;
-
-	default:
-		header('Location: tasks.php');
-		$selected = array_search("Готово", $_POST);
-		$comment = $_POST['comment'];
-		updateTask($selected, $comment);
-		break;		
+if($_POST['changeTask']){
+	$comment = showTask($selected);
+	$h2 = 'Редактировать задачу';
+	require 'header.php';
+	require 'templates/updateTask.php';		
+	require 'footer.html';
+} else if ($_POST['updateTask']){
+	header('Location: tasks.php');
+	$comment = $_POST['comment'];
+	updateTask($selected, $comment);
+} else if ($_POST['deleteTask']){
+	header('Location: tasks.php');
+	deleteTask($selected);
 }
 ?>
